# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Fiction Platform API
  description: OpenAPI 3.0 Specification for Interactive Fiction Platform (DAG Novel System)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /novels:
    get:
      summary: List all novels
      responses:
        '200':
          description: List of novels
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Novel'
    post:
      summary: Create new novel (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                captions:
                  type: string
                  nullable: true
                introduction:
                  type: string
                  nullable: true
                cover_image:
                  type: string
                  nullable: true
                category_ids:
                  type: array
                  items:
                    type: integer
                status:
                  type: string
                is_published:
                  type: boolean
                is_completed:
                  type: boolean
      responses:
        '201':
          description: Novel created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [message, novel_id]
                        properties:
                          message:
                            type: string
                          novel_id:
                            type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /scenes:
    post:
      summary: Create new scene (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, chapter_id, title]
              properties:
                novel_id:
                  type: integer
                chapter_id:
                  type: integer
                title:
                  type: string
                content:
                  type: string
                image_url:
                  type: string
                type:
                  type: string
                  default: normal
                status:
                  type: string
                  default: draft
                ending_type:
                  type: string
                  enum: [Good Ending, Bad Ending, True Ending, Secret Ending]
                ending_title:
                  type: string
                ending_description:
                  type: string
                choices:
                  type: array
                  description: Choices to create after the scene is created
                  items:
                    type: object
                    required: [label, to_scene_id]
                    properties:
                      label:
                        type: string
                      to_scene_id:
                        type: integer
                      choice_id:
                        type: integer
                      targetSubScene:
                        type: string
      responses:
        '201':
          description: Scene created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [message, scene_id]
                        properties:
                          message:
                            type: string
                          scene_id:
                            type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: Invalid input payload
...
    Novel:
      type: object
      properties:
        novel_id:
          type: integer
        title:
          type: string
        captions:
          type: string
        introduction:
          type: string
        cover_image:
          type: string
        status:
          type: string
          enum: [draft, published, suspended]
        is_published:
          type: boolean
        is_completed:
          type: boolean
        is_banned:
          type: boolean
        ban_reason:
          type: string
        author_id:
          type: integer
        author_name:
          type: string
        pen_name:
          type: string
        views:
          type: integer
        chapter_count:
          type: integer
        scene_count:
          type: integer
        like_count:
          type: integer
        bookshelf_count:
          type: integer
...